Create the playGame function
Created the playGame function and the associated displayGameWinner
function. The playGame function calls on the playRound function until
the player or the computer has won 3 times (best of 5). It then calls
the displayGameWinner function to display the overall winner to the
console. The playGame function is called outside of any functions to
simulate the best of 5 game.