Make buttons functional
Added event listeners to the rock, paper and scissors buttons so that
the playRound function is called if they are clicked. The
displayRoundWinner function now displays the winner to the web page
instead of the console. The getPlayerChoice function was removed as the
player choice is now determined by which button they press.