Update displayGameWinner function
Updated the displayGameWinner function so that the winner is displayed
on the web page instead of to the console. The winner now needs to win
five rounds to win the game (previously it was a best of five, i.e.
three points to win). Also removed the playGame function as it is no
longer required.